# Unroll pure go xor loop